requirements.txt
atlassian-python-api>=3.41.1
python-dotenv>=1.0.0
pandas>=2.0.0
